fix some javascript bug
