+++ index.html
@@ -12,12 +12,12 @@
     <header class="navigation">
         <nav class="navbar">
             <img src="images/logo.png" alt="logo" class="logo">
-            <ul class="nav-list">
-                <li id="home" class="nav-link"><a href="#">Home</a></li>
-                <li id="list" class="nav-link"><a href="#">List</a></li>
-                <li id="add" class="nav-link"><a  href="#">Add new</a></li>
-                <li id="contact" class="nav-link"><a href="#">Contact</a></li>
-            </ul>
+            <div class="nav-list">
+                <div id="navbtn-home" class="nav-link"><a  href="#">Home</a></div>
+                <div id="navbtn-list"  class="nav-link"><a  href="#">List</a></div>
+                <div id="navbtn-add" class="nav-link"><a  href="#">Add new</a></div>
+                <div id="navbtn-contact" class="nav-link"><a  href="#">Contact</a></div>
+            </div>
         </nav>
     </header>
 
@@ -69,6 +69,5 @@
     </footer>
     
     <script src="/javascript.js"></script>
-    <script src="/single-page.js"></script>
 </body>
 </html>
